[refactor]: Separate exercise store
