+++ README.md
@@ -9,6 +9,21 @@
 - React-router-dom
 - TawilindCss
 
+### 기술 스택 도입
+
+- react
+  - react 18과 그 이후 기술들, 선언형 UI와 단방향 데이터 흐름들에 대한 이해를 높이기 위해 선택했습니다.
+- vite
+  - webpack은 다양하게 커스터마이징 할 수 있다는 장점이 있지만 세밀한 설정이 필요 없었고, 가볍고 빠르게 개발할 수 있는 vite를 선택하게 되었습니다.
+- redux-toolkit
+  - react가 가지고 있는 단방향 데이터 흐름과 일치하는 flux 패턴 기반의 상태관리 라이브러리로 상태관리를 간소화하고 예측 가능하게 해줍니다.
+- tailwindCSS
+  - 유틸리티 기반의
+- chart.js
+  - 프로젝트에서 대부분 단순한 차트를 표시할 예정이기 때문에, 다양한 커스터마이징이 가능한 D3.js보다는 간소한된 API를 제공하는 chart.js를 선택했습니다.
+- clsx
+  - tailwindCss와 동적 클래스를 많이 사용하는 경우 조건부 클래스들을 쉽게 관리하기 위해 선택했습니다.
+
 ## 프로젝트 구성
 
 - pages > layout > components
@@ -21,25 +36,7 @@
 ## 기술적인 고민
 
 1. input의 공통컴포넌트를 만들 때 controlled component와 uncontrolled component로 나누어 만드는 것은 어떨까?
-2. 에러 관리를 전역적으로 하려면 어떤 방법이 가장 좋을까?
-
-### 기술 스택 도입
-
-- react
-  - react 18과 그 이후 기술들, 선언형 UId와 단방향 데이터 흐름들에 대한 이해를 높이기 위해 선택했습니다.
-- ## vite
-- redux-toolkit
-  - react가 가지고 있는 단방향 데이터 흐름과 일치하는 flux 패턴 기반의 상태관리 라이브러리로 상태관리를 간소화하고 예측 가능하게 해줍니다.
-- tailwindCSS
-  - 유틸리티 기반의
-- chart.js
-  - 프로젝트에서 대부분 단순한 차트를 표시할 예정이기 때문에, 다양한 커스터마이징이 가능한 D3.js보다는 간소한된 API를 제공하는 chart.js를 선택했습니다.
-- clsx
-  - tailwindCss와 동적 클래스를 많이 사용하는 경우 조건부 클래스들을 쉽게 관리하기 위해 선택했습니다.
-
-### 궁금
-
-1. 왜 TailwindCss에는 Autoprefix와 PostCSS가 필요한가?
+   링크(https://ramirami.tistory.com/200)
 
 ## 기능
