+++ src/components/common/AuthGuard.tsx
@@ -1,27 +1,19 @@
 import React, { ReactNode, useEffect } from 'react';
 import { Navigate, useNavigate } from 'react-router-dom';
-import { useSelector, useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
 
 import { setNavigateFunction } from '#/api/axios';
 
-import { getUserInfo } from '#/stores/user/action';
-import { AppDispatch, RootState } from '#/stores/store';
+import { RootState } from '#/stores/store';
 
 const AuthGuard = ({ children }: { children: ReactNode }) => {
   const navigate = useNavigate();
 
-  const dispatch = useDispatch<AppDispatch>();
   const user = useSelector((state: RootState) => state.user);
 
   useEffect(() => {
     setNavigateFunction(navigate);
   }, [navigate]);
-
-  useEffect(() => {
-    if (!user) {
-      dispatch(getUserInfo());
-    }
-  }, [dispatch, user]);
 
   if (
     !user &&
